Fix variable initialization due to jump bypassing it
This fixes the following error in mingw gcc toolchain. Also clang
also have same error.

src/_tkagg.cpp:274:26: note:   crosses initialization of 'bool tk_ok'
  274 |     bool tcl_ok = false, tk_ok = false;
      |                          ^~~~~
src/_tkagg.cpp:274:10: note:   crosses initialization of 'bool tcl_ok'
  274 |     bool tcl_ok = false, tk_ok = false;
      |          ^~~~~~

According to C++ standard (6.7.3):
It is possible to transfer into a block, but not in a way that bypasses
declarations with initialization. A program that jumps from a point where
a variable with automatic storage duration is not in scope to a point
where it is in scope is ill-formed unless the variable has scalar type...