Fix links to rcParams sample matplotlibrc
The anchor ID was changed, but as this was entered manually, it did not
break the build.

Instead, insert a pending cross-reference so that Sphinx will resolve it
and warn if it breaks.

The output changes minimally from `external` to `internal` class, and
also gains `std std-ref` classes due to it being a standard reference.