Make all matplotlib.tri submodules private
Users should access all elements through the outer namespace
matplotlib.tri.

Back-compatibility for the old module names will be added in a separate
commit. If done in the same commit, git would interpret this as
a modified file plus a new file and not as a rename. With the separation
and the rename we keep the history.