Reparametrize offsetbox calculations in terms of bboxes.
Passing a single bbox instead of (xdescent, ydescent, width, height)
separately is easier to follow (see e.g. the changes in VPacker and
HPacker, which no longer have to repeatedly pack/unpack whd_list), and
avoids having to figure out e.g. the sign of the descents and whether
width/height includes the descents, for example.

Currently get_offset keeps a back compatible signature (we *could*
consider killing the old signature but let's not do that for now), and
_get_bbox_and_child_offsets is private because I *may* want to later
also change the convention to make offsets relative to the bbox (0, 0)
point rather than the bbox lower-left corner.